requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.38.0
gunicorn==20.1.0
idna==3.4
ipykernel==6.20.2
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.1
jupyter_core==5.1.5
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
openpyxl==3.1.2
packaging==23.0
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2022.12.19
platformdirs==2.5.4
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
psycopg2-binary==2.9.6
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
requests==2.29.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.8.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
virtualenv==20.16.7
virtualenv-clone==0.5.7
wcwidth==0.2.6
Werkzeug==2.2.2
